requirements.txt
discord.py>=2.0.0
python-dotenv
requests
beautifulsoup4
lxml>=4.9.0
scrapy>=2.11.0 